requirements/install.txt
python-dotenv
tqdm
openai>=1.43.0
numpy>=1.24.4
PyPDF2
click